requirements.txt
python-decouple==3.8
pandas==2.1.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

#for pgcli
pgcli==4.0.1
psycopg-binary==3.1.12